Backend/NeuroNet/requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.4.1
Django==2.2.7
django-extensions==2.2.5
django-floppyforms==1.8.0
django-taggit==1.2.0
djangorestframework==3.11.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
Pygments==2.5.1
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
